Move GameSync to its own namespace and separate out util classes.
