+++ Assets/Scripts/System/NetworkSync.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: adcd8982f740f42e0b0b1c8d67e022d5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
