+++ Assets/Scripts/System/NetworkSync/Game.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd110ae6ff20d43a7ba3818f2127ac4c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
