+++ Assets/Scripts/System/NetworkSync/GameSync.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ba767428e09743d9ac47173ca9bddf7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
