+++ Assets/Scripts/System/NetworkSync/Playlist.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 06db9fd8d6511445dbc1699838433586
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
