+++ Assets/Scripts/System/NetworkSync/SluggedItem.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 039a599164a8f6b419056ab9d3c91d2b
-timeCreated: 1476756731
-licenseType: Free
+guid: da77ebaf8ac284e719a7eecd436aa4a7
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
